refactor(storage): move StorageBackend to lib/artifact_manager; remove pusher shim
Summary:
Intent:
- StorageBackend is a general infrastructure concern reusable across any
  workflow task, not just the pusher — placing it in lib/artifact_manager
  mirrors how ModelRegistryClient lives in lib/model_manager/registry
- No pusher shim: pusher/__init__.py (PR 4) will re-export from
  lib.artifact_manager directly, avoiding unnecessary indirection

Changes:
- Created michelangelo/lib/artifact_manager/ with __init__.py re-export
- Moved storage_backend.py to canonical location lib/artifact_manager/
- Moved tests to lib/artifact_manager/tests/storage_backend_test.py
  (renamed to *_test.py to match lib/model_manager convention)
- Updated test import to michelangelo.lib.artifact_manager.storage_backend

Test Plan:
ruff check michelangelo/lib/artifact_manager/
pytest michelangelo/lib/artifact_manager/tests/
# 8 passed

Revert Plan:
Revert this PR via git revert <commit_sha>.
