+++ python/michelangelo/lib/artifact_manager/storage_backend.py
